Remove useless order-only-prerequisites
They only mean that the target (for example bytecode) does not need
to be updated if the prerequisite was touched. Since the two targets
bytecode and install-bytecode are phony, this doesn't have a effect.
(I think I originally intended to only order the two targets if both
of them are executed anyway, to avoid problems with parallel make...)