Divisio_engraver
In ancient staff contexts, Divisio_engraver creates Divisio grobs in
response to \caesura, \section, \fine, and \repeat volta.  Bar_engraver
is configured not to create visible bar lines for those events.

Divisio grobs are drawn like BreathingSign grobs, but are break-aligned
in the 'staff-bar position like BarLine grobs.

The caesuraType context property tells Divisio_engraver which sign to
to create for an unarticulated \caesura.  The \divisioMinima and
\virgula commands override caesuraType to set their appearance.
\divisioMaior and \divisioMaxima add fermata articulations, which the
caesuraTypeTransform function converts to the appropriate symbols.

caesuraTypeTransform might look like an unnecessary complication in this
commit, but the articulations in the music are appropriate and may be
useful for future MIDI support.  Also, transform functions will have a
role in supporting caesura styles that adapt to bar lines.

convert-ly is "not smart" about this change, but it explains the change
if it finds BreathingSign or Breathing_sign_engraver in the same file as
one of the predefined chant contexts.

Breathing_sign_engraver is still present in voice contexts, so scores
that bypass the divisio functions defined in gregorian.ly and directly
use \breathe with overrides should work as they used to.

KNOWN ISSUES

* Volta brackets do not align to Divisio as they do to BarLine.