Remove rhythmic-head-interface from AmbitusNoteHead
As its name suggests ("rhythmic"), the rhythmic-head-interface is meant
for actual note heads with a duration, not special heads like
AmbitusNoteHead that are merely a symbol in a duration-less notation
element.  In particular, a rhythmic head triggers the creation of a
NoteColumn.  Issue #1113 was caused by a NoteColumn being created at the
first moment without any actual note head in it because the
AmbitusNoteHeads would be acknowledged by Note_column_engraver.

Fixes #1113