Use a unified output-def-kind variable instead of is-{paper,layout,midi}
Although rare, output defs are sometimes written nested inside other
output defs, as in

\paper {
  oddHeaderMarkup =
    \markup \score {
      \layout { ... }
      ...
    }
}

The design of the parser is that scopes are nested, so the \layout in
the example above sees variables coming from \paper.  This wasn't true
in Guile 1.8 due to a Guile bug (see issue #6364 for details), but it
is now that we are running Guile 2.2.

This inheritance does not play well with the way the kind of an output
def (paper/layout/midi) was previously recognized, via variables
is-paper, is-layout and is-midi.  In the example, \layout would
inherit is-paper = ##t from \paper, making the parser wrongly
recognize it.

Independently, it's inelegant to have 3 variables for a single piece
of information (the kind), and it raises questions about whether
is-paper and is-layout can (legitimately) be true at the same time.

So, solve this by unifying these variables in an output-def-kind
variable that can be 'paper, 'layout or 'midi.  I think it would be
good to ultimately make this a member of the Output_def rather than a
variable in its module, as it's an intrinsic property of the output
def, but this will require some reworking of paper.scm.

Fixes #6364