Fix error handling for Scheme errors to display source location info
When reading and evaluating user Scheme code, LilyPond uses Guile's
catch in order to continue compiling even if there is an error in the
code.  The exception data actually does not include the source
location where the error occurred.  Guile displays it for errors not
caught by using a pre-unwind exception handler that prints the message
using the current call stack for backtraces.  We have to mimic this.
About this technique, see
https://www.gnu.org/software/guile/manual/html_node/Pre_002dUnwind-Debugging.html

With a bit of additional work, this even lets LilyPond display the
error on the line of code that triggered it in the LilyPond file, even
if the error occurred deep in Guile/LilyPond functions that the Scheme
code from the .ly file called.

Furthermore, LilyPond's error handling is now consistent with Guile's
in that it prints a backtrace if -ddebug-eval is given.  The backtrace
provides more information, but displaying it is still turned off by
default because it's scary for average users.

Due to all this error handling that is sensibly different from
Guile's, the -dno-protected-scheme-parsing option now also uses catch,
and aborts compilation itself.  This is so that the display of errors
remains the same in this mode.

Fixes #5992