\fine does not end iteration
Let iteration continue after \fine, requiring the use of \volta to
filter out any unwanted music that follows.

Rework Jump_engraver to create Fine text at the written end of the music
and then remove it if it wasn't wanted.

Let convert-ly issue a warning if it finds \fine and a folded repeat.

Closes #6392.