hairpin.cc: add left padding to broken hairpins
Adding left padding to broken hairpins is not only done in some published scores
(see https://gitlab.com/lilypond/lilypond/-/issues/650#note_1014268644 for details)
but it also solves some cases where the hairpin is too high to the left because
one accidental of the key signature pushes it upwards. This commit implements
the suggestion of Mike Solomon.

Closes #650.