Remove safe mode (1): documentation
The short story is that safe mode is utterly broken and virtually
impossible to fix.  Instead of promising security that does not exist,
it is better not to promise any security at all.

The long story has been discussed between the current developers.  As
an exception to the usual rules, this discussion is private because of
its security sensitiveness.

This commit contains the removal of documentation about -dsafe.  The
actual removal from the code is done in several later commits.