Add note columns as side support for trill spanners
A side-positioned grob is normally expected to declare its support in
the side-support-elements array.  TrillSpanner did not do this, and it
worked because outside-staff positioning does a second pass that fixes
the collisions, but adding the support is the intended usage and
should make for more accurate pure estimations.  It also fixes
collisions in case outside-staff-priority is turned off.  Currently,
doing that is necessary to bring the trill into the staff.