Fix grob::when and grob::rhythmic-location on dead spanners
Interfaces are cleared for dead grobs, so a dead spanner would not be
recognized as a spanner and the functions would error out trying to
take the paper column of a spanner.  Instead, use ly:spanner?, which
continues to recognize the grob as spanner even after death because it
checks unsmob<Spanner> (the_grob).