Fix roman page-number-type
The page breaker often needs to do tentative passes to score breaking
configurations. Sometimes, when doing so, it passes 0 as page number.
This would cause an error previously because 0 can't be converted to
roman page numbers.  Instead, use arabic numbers in this case.  A
FIXME is added because I'm not sure if the page breaker really has
good reasons to pass 0 or has an off-by-one error somewhere, but I
think making number-format foolproof is a good idea regardless.

Fixes #6401