Provide option for using compile to run embedded Scheme code
Guile 2+ has two ways of running arbitrary code from data: evaluation,
and compilation.  The latter has a compilation overhead, but
evaluation makes for very unhelpful error messages.  This adds a
-dcompile-scheme-code option that lets LilyPond use the compiler.
This is not the default for the time being, and for two
reasons. First, due to the way the compiler interacts with BDWGC, it
does not support compiling more than a few thousand expressions [1].
This makes it impossible to use for 'make check' until that problem is
fixed in Guile.  It seems unwise to make something the default for
users if it can't be tested in 'make check'.  Second, the compiler has
some incompatibilities with the evaluator, which would make it better
to switch to it at an early point in the release cycle, while we're
just two weeks away from branching 2.24.

[1] https://lists.gnu.org/archive/html/guile-devel/2022-08/msg00033.html