Fix mutation in scheme-text-spanner.ly
According to Scheme standards, it is an error to attempt to mutate a
"literal constant".  Guile enforces this restriction for constant
parts of quasiquotes -- but not only when byte-compiling, not when
evaluating.  Without this fix, the regression test
scheme-text-spanner.ly would error out with the upcoming switch to
compilation for user Scheme code because it does this kind of
mutation.