Introduce StaffSymbol.widened-extent property
The widened-extent of a StaffSymbol is its Y extent, possibly widened
if it's too short for a bar line.  The BarLine code used to compute
this itself, but now it reads it in StaffSymbol, which will allow
reusing the same property for highlights in the next commit.

This entails a behavior change: now, the correction that makes the bar
line extend to the middle of the extremal staff line rather than to
the end is also applied in the case where the bar line actually ends
on a "virtual" staff line, when widened-extent is wider than Y-extent.
The difference is insignificant.