Add staff highlights
Staff highlights are a way to color the staff symbol during a musical
passage for analytical and pedagogical purposes.  This implementation
is adapted from a snippet which (to my surprise) gained some
popularity, originally posted at
https://lists.gnu.org/archive/html/lilypond-user-fr/2021-12/msg00038.html
and is the subject of the blog post (in French)
https://myrealbook.vintherine.org/un-peu-de-couleur.html

See also
https://gitlab.com/lilypond/lilypond/-/merge_requests/1545#note_1061650038
for a discussion of related features that could be added in the
future.

I've considered single-moment staff highlights, but decided not to
support them for now because event order is unspecified, which yields
ambiguity:

  %% One of these starts a single-moment highlight stopped immediately,
  %% the other starts a highlight that continues afterwards.  Who's who?
  \staffHighlight "red"
  \staffHighlight "green"
  \stopStaffHighlight

Known limitation: if a staff highlight ends on a repeat bar line, it
includes it, which is most likely unwanted.  This is because
highlights need to extend to the right of the bar line on their right
bound, otherwise it there is an ugly discontinuity at the end of a
system when a highlight spans several staves that aren't connected by
span bars.  When considering bar lines like \bar ":|.|:", it's clear
that solving this would need a way for each bar type to tell where its
"center" is.  That is left as a TODO in this commit.