Make flageolet glyph smaller
See https://lists.gnu.org/archive/html/lilypond-devel/2022-08/msg00021.html
for an analysis of the flageolet symbol with some examples of published
scores.  So far the documentation not only shows how to make a flageolet
but also how to make it smaller.  This commit makes the flageolet symbol
smaller by default and deletes mentions about making it smaller.
Additionally, the glyph is now slightly thicker, following the overall
heavy character of Emmentaler (thanks to Kieren for the suggestion!).
It follows the general recommendation of Kurt Stone against big flageolets.