Fix to include config.hh first
This commit puts config.hh first in the include order
for source files that include it.

autoconf `AC_USE_SYSTEM_EXTENSIONS` defines
a feature macro `_GNU_SOURCE` in `config.hh` as needed.
The feature macro can be defined
before the inclusion of any system header files
to obtain the necessary function declarations.

On Cygwin environment, a compile error was raised
that the function is not declared
because the system header file is included before config.hh.
https://lists.gnu.org/archive/html/bug-lilypond/2022-08/msg00008.html

See
https://gitlab.com/lilypond/lilypond/-/merge_requests/1554#note_1065324962