Swap stencils for bassoon woodwind-diagram keys one and six
The cc-one key was given an elliptical stencil, this is in fact a hole and the
cc-six key should in fact be the elliptical key. The right hand bes key position
is adjusted to fit better between five and six. See resources at the
International Double Reed Society website https://idrs.org/resources/BSNFING/HEKEYPIC.HTM