Doc: Remove tocFormatMarkup = #markup from an example
This is useless because tocFormatMarkup is used by
format-toc-toplevel, which is not used in tocItemMarkup as redefined
in the example, and it causes an error because markup is a macro and
writing the name of a macro as an expression is not valid in Guile 2
(more precisely, it's valid if the macro defines a pattern for
invocation as "identifier syntax", but it then causes macro expansion,
it doesn't give the macro object itself).  This went unnoticed without
breaking the doc build because of issue #6344 (“Scheme errors no
longer result in a non-zero return code in 2.22”).