Add Caesura_engraver for modern caesura engraving
Depending on the result of passing the value of caesuraType through
caesuraTypeTransform, this engraver may create a BreathingSign with
CaesuraScript grobs aligned to it, or it may create CaesuraScript grobs
aligned over a BarLine.

If this engraver observes a BarLine, it calls caesuraTypeTransform again
with the new information, and if necessary, recreates its grobs.

New Grobs
* CaesuraScript: outside-staff comma, fermata, etc.

New Engravers
* Caesura_engraver: make staff-scope BreathingSign and CaesuraScript
* Non_musical_script_column_engraver: stack CaesuraScript grobs

Wish list
* a more bulbous comma (issue 6343)

Limitations
* Scripts in the 'scripts entry are tweaked jointly with
      \tweak ... \caesura
  whereas articulations following \caesura may be tweaked individually.
* caesuraTypeTransform can add scripts by adding to the 'scripts entry,
  but not by adding to the 'articulations entry.

Supported styles
* breath in mid measure, script at bar line (e.g. caesura→fermata)
* script in mid measure, other script at bar line (e.g. comma→fermata)
* remove mark or script at bar line, but keep articulations
* etc.