Prevent Instrument_name_engraver from being consisted several times in regtests
This fixes two cases of duplicate engravers that were spotted while
testing a change that enforces a single instance of each engraver per
context.  It's harmless strictly speaking, but leads to two instrument
names overprinted.

In multi-measure-rest-instr-name.ly, don't re-consist
Instrument_name_engraver, as it's already in Staff by default.

In filter-translators.ly, remove Instrument_name_engraver in a \layout
first, so that the test actually tests it's being added by \with in
layout output.