Replace (ice-9 curried-definitions) with a fixed version
Guile's (ice-9 curried-definitions) module is buggy with docstrings:
in a curried definition, the docstring is placed in the innermost
lambda, making it visible on the result of applying the curried
function until the last lambda, but not on the curried function
itself.  This has been fixed for the to-be-released Guile 3.0.9 with
the patch on https://debbugs.gnu.org/cgi/bugreport.cgi?bug=50068 Until
Guile 3.0.9 is released for long enough and works well enough for us
to require it, use a new module (lily curried-definition) with the
fixed version of the code, so we get correct docstrings for curried
functions in the Internals Reference.