Add more optional fallback parameters to Scheme functions
LilyPond's Scheme interfaces often use '() as fallback for missing
values, which is unfortunate since the usual way in Scheme is to use
#f.  '() is inconsistent with Guile functions, and does not allow idioms
idioms like and, or, and-let*.  This flaw is difficult to fix, since
the functions have been around for so long, but it can be alleviated
by adding optional parameters specifying the fallback explicitly.
Most property accessors have had this for a long time, and
ly:grob-object received it in 373d16a11f.  This adds such fallbacks
for these functions: ly:context-property-where-defined,
ly:grob-parent, ly:output-def-parent, ly:spanner-bound.