Prepare Changes document for release
The structure is again based on the NR, with sections "New for musical
notation" and "New for specialist notation", that have subsections as
in the NR (pitches, rhythms, ..., ancient notation, world music). There
is a section named "Miscellaneous improvements" for the rest, either
related to the remaining chapters in the NR or internals, for example
improvements for Scheme translators or the Grob infrastructure.

Compared to the last stable release, the document now uses nodes and
unnumbered sections that add entries to the Table of Contents on the
left for easier navigation on the single page.