Fix ly:position-on-line?
It was returning the negation of the correct result when called on a
non-StaffSymbol grob.  The reason is that it would call
Staff_symbol::on_line instead of Staff_symbol_referencer::on_line, and
thus try to look up line-positions and such not on the staff symbol
but on the argument grob itself.  Then the (only) call site in
flag-styles.scm was doing

  (if (ly:position-on-line? stem-grob stem-end) "1" "0")

in contrast to the C++

  Staff_symbol_referencer::on_line (stem, p) ? "0" : "1";

which inverted the result again.