Match X extent of Dots to real extent
Before, the Dots stencil had an extra X extent at the left to pad the
dots from the note heads.  There are several problems with this.  One
is that it's surprising, since X-extent normally represents the width
of the object itself.  A more immediate problem is that slurs were
being pushed away further than necessary when starting or ending on
dotted notes, because Slur_configuration::score_extra_encompass in
slur-configuration.cc has a special path for grobs that overlap with
the note heads, and this path was being triggered for dots, while they
are in fact not overlapping horizontally with the heads.  This code
path would cause the slur to push itself away enough to start/end at a
higher point than the top of the dot.

This fixes these by moving the padding from Dots to DotColumn.  The
new property DotColumn.padding determines how much the DotColumn moves
away from the note heads.

This also has the nice consequence of making DotColumn.padding
configurable; before, there was no simple way to configure the
distance from heads to dots.

Fixes #3863