binaries: Disable win32 Dlls in bdwgc
For lack of a better method for finding static data sections in shared
libraries, bdwgc on Windows scans all memory pages and (temporarily)
adds them to the root set. This leads to problems and sporadic crashes
if a memory page disappears while marking. Until a proper solution is
found, disable this entire mechanism because by design of statically
linking both bdwgc and libguile, we can guarantee that all root sets
relevant for garbage collection are in the data section of the main
executable, which can be conveniently located by looking at a single
static variable and determining the surrounding pages.

Should fix #6361