Change convention for ly:skylines-for-stencil axis argument
In hindsight, the previous choice of convention, namely passing X to
get horizontal skylines, was a mistake from b866f652a3.  Now that more
Scheme interfaces to skylines need to be added, using a different
convention in the interface than in internals isn't tenable,
especially for adding ly:make-skyline, where "horizon axis"
corresponds to the intuition of how the skyline is built, even though
this does mean that (ly:make-skyline ... X) creates what is called a
"vertical skyline".

ly:skylines-for-stencil was introduced in 2.23.9, so it should be fine
to change it in this way.