Make StaffSymbol.line-positions always set
Before, line-positions was only for user overrides.  All places accessing it
needed to fall back on using line-count if it was empty.  Now, the default for
line-positions is a callback that computes the usual line positions from
line-count, obviating those fallbacks.

This entails a minor change: before, \override Staff.StaffSymbol.line-positions
= #'() would specify to revert to the behavior of computing staff lines from
line-count, but now it means to print no staff line, like line-count = 0 (which
makes more sense).  convert-ly takes care of this.