Restate convert-ly rule for \override syntax in 2.23.11
While \override and \revert syntax was introduced in 2.17.6, a warning was added
for the old syntax in 2.23.11 with aee38e53e39116b0cd30c3996f24ba3efb296c0e,
which for a number of people is the first time they realize some of their
overrides are using old syntax, and some copied things from snippets found on
the net or mailing list answers, which bypasses convert-ly.  Therefore, this
restates the convert-ly rule that was used in 2.17.6 for 2.23.11, so people who
run convert-ly won't ever see this warning.

(cherry picked from commit 03855a333fbfe7b2550039afae28aeead16d4c94)