max-slope-factor for tuplet brackets
This commit limits the slope that a tuplet bracket can have in
some circumstances.  The maximum slope of a tuplet bracket is so:

1) Proportional to the slope of the beam, when the beam spans the
whole duration of the tuplet (default Lilypond behavior).

2) Proportional to the slope of the beam when the bracket is
parallel to a beam that spans the whole duration of the tuplet
(i.e. when the tuplet bracket is on the note head side) and the
slope of the bracket is steep.  In a sequence of discontinuous
beams, the last beam is considered (as optical, directional
guide).

3) Otherwise, it is limited to the maximum slope calculated, which
can be changed with the new property `max-slope-factor`.

Edit: thanks to Jean the new type predicate `non-negative-number?`
has been added.

Closes #4928
Closes #992