Snippet: changes to controlling-tuplet-bracket-visibility
Apparently the bracket visibility property #'if-no-beam was the
default in the past.  The wording of the snippet still implied
that it was the default in contradiction with the default behavior
described at the beginning of the snippet.  This commit corrects
that and makes use of `\textMark`.

(cherry picked from commit f88270ab60ac9383a5ac97349cacc63b89b82107)