Don't print tuplet bracket when tuplet number is wider
This was already part of the code.  However, because `scm_is_true`
yields true unless it's false, the condition that turned the
bracket visibility off when the tuplet number is wider than the
tuplet bracket was never true in this case, and as consequence a
tuplet bracket was still being printed even when it shouldn't,
resulting in negative lengths (i.e. "inside-out brackets").

The factor 0.75 was unnecessarily magnifying the length of the
gap as a workaround, but the truth is that the condition
`!bracket_prop` never yielded true (`bracket_prop` would've had
to be false in the first place, but then `bracket_visibility` was
also false), so there was no way of testing the delay in the
printing of the tuplet bracket that 0.75 was introducing.  This
patch removes it instead.  It is sufficient that the tuplet
bracket has a positive value larger than the tuplet number + 1
(the gap) to be printed.

Closes #3901