lilypond-book syntax: fix `lilypond` environment option usage
Proper lilypond environment syntax is

    \begin{lilypond}[options ...]

rather than

    \begin[options ...]{lilypond}

and the latter is only supported out of compatibility reasons.  The correct
syntax is the one that would be required for actual LaTeX environments, will be
the only one supported by syntax highlighting of LaTeX editors, and would be
required for an implementation working in LaTeX proper rather being preprocessed
by lilypond-book.