Simplify and unify connect-to-neighbor callbacks
These were all the same and were duplicating the logic of
Spanner::broken_neighbor.  Unify them into a single callback attached to
spanner-interface that reuses Spanner::broken_neighbor.

Note that ly:tuplet-bracket::calc-connect-to-neighbor was triggering 'positions,
with a comment saying that this could lead to suicide, but there is nothing in
the current code of the grobs that were using
ly:tuplet-bracket::calc-connect-to-neighbor that could trigger suicide in
'positions.