Clean up extract_item_set
Instead of mapping non-Items to null pointers, filter them out and issue
a programming error.  Coherent_ligature_engraver ignored null pointers
anyway, and other callers did not bother to check for them.

Stop using extract_item_set in Extender_engraver and Hyphen_engraver.
There isn't much value in checking the full array when only one element
is used.

Implement the internal function as a template so that it can easily be
instantiated for Spanner, if necessary.