Vertically align tuplet numbers of flat brackets
Issue #5615 happened because the `offset` of flat tuplet brackets
was being set too strictly (to `dy = 0` to be able to be
considered flat) whereas optically they look flat even with values
slightly different than 0.  This led to an optically wrong
collision detection and rounding issues which gave the impression
that tuplet brackets don't vertically align.

Closes #5615