Create FingerGlideSpanner from finger glide event
Use the finger glide event as cause for FingerGlideSpanner instead of the note
event.  This makes point-and-click point from the FingerGlideSpanner to \glide
in the input, not to the note it's attached to.  It also avoids the need for
Finger_glide_engraver to reimplement Tweak_engraver's logic, which was
incomplete.

As a consequence, "\tweak color c-1\glide" no longer prints a red
FingerGlideSpanner.  This is just as expected (e.g., the stem or the fingering
don't become red either).

Fixes #6292