Web: Update Download part for 2.24
Remove installation instructions that were relevant for binaries produced by
GUB.  Just link to the learning manual, where the installation tutorial lives
now.  This makes the per-OS pages very short, so merge them into a single page.

Update create-weblinks-itexi.py to create only links to GitLab, the new storage
for releases.

Also add a few links relevant for users who upgrade (as opposed to downloading
LilyPond for the first time), mainly to the Changes document.

The French translation is updated (needs to be in this commit because otherwise
the use of removed macros would break the build).

Fixes #6458