hyphenate-internal-words.scm: Minor improvement
Instead of having calls to `@hyphenation` for each entry, use a single call
for all entries.

Besides more efficient, this avoids a crash with `luatex` ID 7509 (the
binary part of TeXLive 2022): it cannot handle a large amount of
`@hyphenation` calls due to an allocation bug.

Also sort entries to get identical results for different runs.