doc: Fix index entries 2/2
Currently, the `texindex` program (used by `texinfo` to sort and unify index
entries) has no support for locale-specific collation.  As a consequence,
index entries starting with a non-ASCII character are sorted more or less
arbitrarily before letter 'a'.

By replacing such characters with a base character preceded by a texinfo
accent macro, this can be alleviated; `texindex` then uses the base
character as a sorting key.

The following script (for GNU sed) was used to do the replacement.

```
for f in `find . -name '*.itely'`; do
  sed -i \
      -e 's/@cindex Ä/@cindex @"A/' \
      -e "s/@cindex á/@cindex @'a/" \
      -e 's/@cindex à/@cindex @`a/' \
      -e 's/@cindex Č/@cindex @v{C}/' \
      -e 's/@cindex č/@cindex @v{c}/' \
      -e "s/@cindex É/@cindex @'E/" \
      -e "s/@cindex é/@cindex @'e/" \
      -e "s/@cindex í/@cindex @'i/" \
      -e 's/@cindex ö/@cindex @"o/' \
      -e "s/@cindex ó/@cindex @'o/" \
      -e 's/@cindex Ř/@cindex @v{R}/' \
      -e 's/@cindex ř/@cindex @v{r}/' \
      -e 's/@cindex Ü/@cindex @"U/' \
      -e 's/@cindex ü/@cindex @"u/' \
      -e "s/@cindex Ú/@cindex @'U/" \
      -e "s/@cindex ú/@cindex @'u/" \
      -e 's/@cindex ù/@cindex @`u/' \
      $f
done
```