Fix direction for whole note tremolos on third line
When a whole note tremolo began on the third line of the staff it
produced a warning cascade:
— `programming error: cyclic dependency: calculation-in-progress
encountered for Stem.direction`
— `direction of grob Stem must be UP or DOWN; using UP`
— `weird stem size, check for narrow beams`...

These warnings were there because the beam could be left without a
direction (i.e. direction "CENTER", which is 0) when the first
note didn't have a "normal stem" (i.e. when it was a whole note)
and the note was on the third line.  At the same time, though,
`ly:stem::calc-direction` was relying on the direction of the
detected beam to set the direction of the stem (which for whole
notes are invisible, but have a direction).  This of course led to
a cyclic dependency which this patch fixes.  As a consequence of
this fix now the tremolo direction of whole note tremolos is
similar to that of beamed tremolos: instead of producing in the
end a "random" direction, now whole note tremolos use the default
beam direction if the default direction of the stem is still
CENTER by the time it's called by `ly:beam::calc-direction`.