Whole note tremolos (and longer) don't collide with accidentals
The "gap" property is now internally converted to a
Drul_array<Real> type.  This makes it possible to adjust the right
gap to make room for accidentals.  To compensate for shorter
tremolos caused by a wider right gap, a `minimum-length` property
has been introduced that gets activated only in the presence of a
"whole note tremolo" by using the new Scheme callback
`ly:beam::tremolo-springs-and-rods`.

Closes #704
Closes #3143