mf: Use more precise value for font revision
According to

  https://lists.freedesktop.org/archives/fontconfig/2022-December/006927.html

the font revision field in the 'head' table of an SFNT font gets used to
distinguish between fonts that FontConfig does consider identical otherwise.

Up to now, we relied on FontForge to automatically fill this field.
However, for a version 'XX.YY.ZZ', the revision field only contained
'XX.YY'; as a result, FontConfig considered fonts from 2.23.5 as the same as
from 2.23.82.

We now set the value manually to 'XX.YYZZ'.