Accept bare pitches/durations as music arguments to markup commands
This adds a tiny bit of the smart argument handling music functions have to
markup commands.  Namely, if a command expects a Scheme argument (i.e., the
predicate is not markup? or markup-list?), and { <pitch> } or { <duration> } is
passed, the music interpretation as note event of those constructs is tried as
well against the predicate.  The most obvious use case is

  \markup \rhythm { 8. }