Replace rhythmic-head-interface with note-head-interface for TrillPitchHead
TrillPitchHead uses ly:note-head::print, therefore it makes sense for it
to have note-head-interface.  On the other hand, it doesn't have a
rhythmic value, so it shouldn't have rhythmic-head-interface.

I have not found any cases making sense musically where this change
would make a significant difference, but it makes all engravers that
acknowledge rhythmic-head-interface easier to reason about.  These cases
not making a lot of musical sense can be mentioned:

    {
      \override DynamicLineSpanner.padding = 10
      \pitchedTrill <c' c''>2^\startTrillSpan_\< dis, d'\!
    }

Now a DynamicLineSpanner will only avoid TrillPitchHead by means of
outside-staff positioning, not by side-positioning, so padding does not
take TrillPitchHead into account anymore.  A trill pitch head is
normally close to a normal head, so this is unlikely to make a big
difference in practice.  Said differently, LilyPond sets different
paddings for "main" expected collision types (with note heads) and
"extra" collisions, and a collision between dynamic and trill pitch head
is now considered an "extra" collision.

    {
      \pitchedTrill
        <c' c''>2^\startTrillSpan-\tweak direction #RIGHT \arpeggio
        d'
      d'
    }


Now the arpeggio doesn't have the trill head as support anymore, so it
is placed on its left.

    <<
      \new Staff {
        \set followVoice = ##t
        \pitchedTrill c'\startTrillSpan d'
        \change Staff = down
        c'
      }
      \new Staff = down {
        s1
      }
    >>

Now a voice follower will only attach to a main head (but it's probably
best not to use a voice follower in this case?).