doc: Add typographically enhanced version of `@code`
If the LilyPond documentation is compiled with LuaTeX, enhance Texinfo's
`@code` macro with some Lua code (and the necessary Texinfo wrapper code) to
do the following things.

(1) Add possible line breaks after `-` and `_`.
(2) Avoid that single-character words at the start or the end of the `@code`
    argument are positioned at the end or start of an output line,
    respectively.  A typical case with LilyPond code is `@code{@{ ...  @}}`.

There must be at least two characters at the start or the end of a word
before a hyphenation point or a possible line break gets inserted.  If a
word already contains manually inserted discretionaries (`@-`) or penalties
(`@/`), it will be ignored.