Make Parentheses.X-extent stencil-based
Parentheses.X-extent defaulted to '(0 . 0), which was done in commit
8f9554e6 as a bugfix for parentheses on breath marks.  However, the
comment was wrong: an empty extent would have worked too, and the
default extent is not empty, but stencil-based (see the Grob
constructor).  The root cause of the problem was the parent of the
parentheses, which was PaperColumn, while it should have been either
BreathingSign or NonMusicalPaperColumn since a breathing sign is
non-musical.  Computing an extent between columns before line breaking
is bound to produce problems.  The parents were fixed in af931c4b1e, so
removing this code is safe.  It fixes cropping, and will help for making
parentheses reserve space.

Fixes #6272