Update Welcome_to_LilyPond.ly
It is referring to LilyPad, which is gone.  Remove it, as it is no
longer used in LilyPad.  Replace it with a file
"internal-backend-test.ly" to keep doing the basic backend exercise in
"make test"; ideally this should be moved to input/regression/, but I
don't understand the build system well enough to do it.