Refactor "gulp file" functions
Make ly:find-file take an optional "strict" parameter.  If #t, ly:find-file
errors out when the file is not found, instead of returning #f.  This allows to
reimplement ly:gulp-file and ly:gulp-file-utf8 more simply, in Scheme and with
Guile APIs.  The end goal is to allow easily reading files in binary while
outputting nice errors if they're not found.

Known difference: ly:gulp-file-utf8 with a size parameter now reads the given
number of characters, not bytes.  Given that Guile strings think in characters,
this makes more sense, and there was not call site of ly:gulp-file-utf8 with a
size.