Documentation es: small change
Perhaps is "conminar" a verb used in this context in some
countries http://lilypond.org/development.es.html.  To me it
sounds quite harsh.  The RAE confirms this (its first meaning is
"amenazar" (to menace)).  Although it's true that people with no
Lilypond experience should rather go to the stable version, it's
not necessary to menace anyone ;-).

Changed to "pedir encarecidamente" following the German version
("ausdrücklich empfehlen").