Cairo: support PDF outlines
This mimics the existing implementation in framework-ps.scm.

The TOC infrastructure is extended to make each TOC entry contain the ID
of its parent, since that is how Cairo allows to define the nesting,
unlike PostScript's pdfmark, which expects each parent to announce its
number of children.

Just like PDF metadata, this works in Cairo ≥ 1.16.

Fixes #6501