Fix inclusion of Texinfo sections into collated-files
During the removal of the stepmake/ directory in commit 96bfa0bccd,
I merged the generic templates into the make/ directory and then
unified and sorted the TEMPLATES into one variable during commit
090578693. As a result, the lysdoc template is now included before
the texinfo template, but the latter reset TEXINFO_SOURCES instead
of appending to it. This made the Texinfo sections disappear from
the collated-files.list and consequently the built HTML and PDFs.

As a fix, define COLLATED_FILES locally in the directories that use
the lysdoc template - they know best which extensions to look for.
This also opens the door for future improvements, such as moving some
of the many files in input/regression/ into subdirectories while
still compiling in a single invocation, and also integrating back
the tests in other/ while adding a note that the visual output is
not considered important.

(cherry picked from commit 906cc6d17b11945b975639d7245cba8fc9f849e0)